renamed application folder and updated imports
