+++ swarm_intelligence_app/.gitignore
