+++ swarm_intelligence_app/common/errors.py
