+++ swarm_intelligence_app/models/models.py
